+++ CHANGELOG.md
@@ -11,3 +11,4 @@ All notable changes to `cartola-fc-api` will be documented in this file.
 ## 0.2.0 - 2022-12-31
 
 - Added Classe with team endpoints
+- Added Classe to transform api response